Remove QualitySettings shadow related value tweaks, not needed anymore.
The HDRP code that was tweaking the values in order to set shadows to "on" was making the QualitySettings asset be always checked out in Perforce. This is because during the tweak process it was actually modifying the asset!

Originally the code was added in PR #2298, but in current Unity versions that is not needed, since <ono-url-prefix>/unity/unity/pull-request/84140/_/lighting/srp-shadow-enable made it so that culling ignores quality settings values when SRP is active. That landed to trunk on 2019 March 14.

I manually checked if overriding quality settings values to "shadows: none" still makes shadows work fine in HDRP, and it does.

This fix is part of fogbugz case 1151097.